+++ skel/share/defaults/ftp.properties
@@ -285,6 +285,17 @@ ftp.loginbroker.family.gsi=gsiftp
 ftp.loginbroker.version.gsi=1.0.0
 ftp.loginbroker.family.kerberos=gkftp
 ftp.loginbroker.version.kerberos=1.0.0
+
+##  This property is a space-separated list of hostnames or IP
+##  addresses to publish for this door.  Hostnames are resolved to an
+##  IP address on start-up, but not subsequently.  Non-wildcard IP
+##  addresses are resolved to canonical name on start-up, but not
+##  subsequently.  Wildcard IP addresses (0.0.0.0 or ::) are replaced
+##  by a dynamically discovered list of IP addresses, based on which
+##  network interfaces are "up".  These are resolved, with the
+##  corresponding hostnames cached for a configurable period (see
+##  'networkaddress.cache.ttl' system property).  An empty value is
+##  equivalent to "0.0.0.0".
 ftp.loginbroker.address = ${ftp.net.listen}
 ftp.loginbroker.port = ${ftp.net.port}
 ftp.loginbroker.root = ${ftp.root}
