+++ skel/share/defaults/webdav.properties
@@ -135,6 +135,17 @@ webdav.loginbroker.version=1.1
 webdav.loginbroker.family=${webdav.loginbroker.family-${webdav.authn.protocol}}
 (immutable)webdav.loginbroker.family-http=http
 (immutable)webdav.loginbroker.family-https=https
+
+##  This property is a space-separated list of hostnames or IP
+##  addresses to publish for this door.  Hostnames are resolved to an
+##  IP address on start-up, but not subsequently.  Non-wildcard IP
+##  addresses are resolved to canonical name on start-up, but not
+##  subsequently.  Wildcard IP addresses (0.0.0.0 or ::) are replaced
+##  by a dynamically discovered list of IP addresses, based on which
+##  network interfaces are "up".  These are resolved, with the
+##  corresponding hostnames cached for a configurable period (see
+##  'networkaddress.cache.ttl' system property).  An empty value is
+##  equivalent to "0.0.0.0".
 webdav.loginbroker.address = ${webdav.net.listen}
 webdav.loginbroker.port = ${webdav.net.port}
 webdav.loginbroker.root = ${webdav.root}
