+++ skel/share/defaults/xrootd.properties
@@ -103,6 +103,17 @@ xrootd.loginbroker.update-threshold=${dcache.loginbroker.update-threshold}
 xrootd.loginbroker.version=2.4
 xrootd.loginbroker.family=root
 xrootd.loginbroker.root = ${xrootd.root}
+
+##  This property is a space-separated list of hostnames or IP
+##  addresses to publish for this door.  Hostnames are resolved to an
+##  IP address on start-up, but not subsequently.  Non-wildcard IP
+##  addresses are resolved to canonical name on start-up, but not
+##  subsequently.  Wildcard IP addresses (0.0.0.0 or ::) are replaced
+##  by a dynamically discovered list of IP addresses, based on which
+##  network interfaces are "up".  These are resolved, with the
+##  corresponding hostnames cached for a configurable period (see
+##  'networkaddress.cache.ttl' system property).  An empty value is
+##  equivalent to "0.0.0.0".
 xrootd.loginbroker.address = ${xrootd.net.listen}
 xrootd.loginbroker.port = ${xrootd.net.port}
